Fixed bugs For Version 6(1.6)
