+++ app/build.gradle
@@ -11,8 +11,8 @@ android {
         applicationId "com.dbot.client"
         minSdk 21
         targetSdk 31
-        versionCode 5
-        versionName "1.1"
+        versionCode 6
+        versionName "1.6"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
