+++ app/src/main/java/com/dbot/client/common/city/SaveCity.java
@@ -1,0 +1,5 @@
+package com.dbot.client.common.city;
+
+public interface SaveCity {
+    void cityData(CityData cityData);
+}
