+++ app/src/main/java/com/dbot/client/login/LoginViewModel.java
@@ -6,8 +6,8 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-import com.dbot.client.login.city.CityData;
-import com.dbot.client.login.city.CityResponse;
+import com.dbot.client.common.city.CityData;
+import com.dbot.client.common.city.CityResponse;
 import com.dbot.client.login.model.LoginResponse;
 import com.dbot.client.login.model.SignUpResponse;
 import com.dbot.client.login.model.User;
