+++ app/src/main/java/com/dbot/client/login/SignupActivity.java
@@ -1,20 +1,13 @@
 package com.dbot.client.login;
 
 import static com.dbot.client.common.CommonFunctions.findCityPosition;
+import static com.dbot.client.common.Popup.SearchCity;
 
-import android.app.Dialog;
-import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
 import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
-import android.text.Editable;
-import android.text.TextWatcher;
 import android.util.Log;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.EditText;
-import android.widget.ListView;
 import android.widget.RadioButton;
 import android.widget.Toast;
 
@@ -25,15 +18,17 @@ import androidx.lifecycle.ViewModelProvider;
 import com.dbot.client.R;
 import com.dbot.client.common.Popup;
 import com.dbot.client.common.SessionManager;
+import com.dbot.client.common.city.SaveCity;
 import com.dbot.client.databinding.ActivitySignupBinding;
-import com.dbot.client.login.city.CityAdapter;
-import com.dbot.client.login.city.CityData;
+import com.dbot.client.common.city.CityAdapter;
+import com.dbot.client.common.city.CityData;
 import com.dbot.client.login.model.SignUpResponse;
 import com.dbot.client.login.model.User;
+import com.google.gson.GsonBuilder;
 
 import java.util.List;
 
-public class SignupActivity extends AppCompatActivity {
+public class SignupActivity extends AppCompatActivity implements SaveCity {
 
     ActivitySignupBinding binding;
     LoginViewModel loginViewModel;
@@ -42,6 +37,8 @@ public class SignupActivity extends AppCompatActivity {
     SessionManager sessionManager;
     List<CityData> cityDataList;
     CityAdapter cityAdapter;
+    SaveCity saveCity;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -76,7 +73,7 @@ public class SignupActivity extends AppCompatActivity {
                 int position = findCityPosition(cityDataList, sessionManager.getCity());
                 binding.tvCitySearch.setText(cityDataList.get(position).getCityName());
                 //binding.spCity.setSelection(position);
-               // binding.spCity.setOnItemSelectedListener(SignupActivity.this);
+                // binding.spCity.setOnItemSelectedListener(SignupActivity.this);
             }
         });
         loginViewModel.getSignUpResult().observe(this, new Observer<SignUpResponse>() {
@@ -123,93 +120,24 @@ public class SignupActivity extends AppCompatActivity {
         binding.tvCitySearch.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                SearchCity();
+                saveCity = SignupActivity.this;
+                SearchCity(SignupActivity.this,cityDataList,saveCity);
             }
         });
 
     }
 
-    private void SearchCity() {
-        // Initialize dialog
-        Dialog dialog=new Dialog(SignupActivity.this);
-
-        // set custom dialog
-        dialog.setContentView(R.layout.dialog_city_search);
-
-        // set custom height and width
-        dialog.getWindow().setLayout(650,800);
-
-        // set transparent background
-        dialog.getWindow().setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
-
-        // show dialog
-        dialog.show();
-
-        // Initialize and assign variable
-        EditText editText=dialog.findViewById(R.id.edit_text);
-        ListView listView=dialog.findViewById(R.id.list_view);
-
-        // Initialize array adapter
-        //ArrayAdapter<String> adapter=new ArrayAdapter<>(SignupActivity.this, android.R.layout.simple_list_item_1,arrayList);
-        cityAdapter = new CityAdapter(SignupActivity.this,getApplicationContext(), cityDataList);
-        // set adapter
-        listView.setAdapter(cityAdapter);
-        //listView.setTextFilterEnabled(true);
-        //listView.setOnItemSelectedListener(SignupActivity.this);
-        editText.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-                Log.d("searchText",s.toString());
-                cityAdapter.getFilter().filter(s.toString());
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-
-            }
-        });
-
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                // when item selected from list
-                // set selected item on textView
-                Log.d("position",String.valueOf(position));
-                if (cityDataList.get(position).getWorkingCity().equals("1")) {
-                    city = String.valueOf(cityDataList.get(position).getId());
-
-                }
-                else {
-                    Popup popup = new Popup();
-                    popup.showCityNotAvailablePopupWindow(binding.tvCitySearch);
-                    city = String.valueOf(cityDataList.get(position).getId());
-                }
-                binding.tvCitySearch.setText(cityDataList.get(position).getCityName());
-
-                // Dismiss dialog
-                dialog.dismiss();
-            }
-        });
-    }
-
-  /*  @Override
-    public void onItemSelected(AdapterView<?> adapterView, View view, int i, long l) {
-        //Toast.makeText(SignupActivity.this,cityDataList.get(i).getId(),Toast.LENGTH_SHORT).show();
-        if (cityDataList.get(i).getWorkingCity().equals("1"))
-            city = String.valueOf(cityDataList.get(i).getId());
-        else {
-            Popup popup = new Popup();
-            popup.showCityNotAvailablePopupWindow(binding.spCity);
-            city = String.valueOf(cityDataList.get(i).getId());
-        }
-    }
 
     @Override
-    public void onNothingSelected(AdapterView<?> adapterView) {
-    }*/
+    public void cityData(CityData cityData) {
+        //CityData cityData = (CityData) parent.getItemAtPosition(position);
+        Log.d("cityData", new GsonBuilder().setPrettyPrinting().create().toJson(cityData));
+        if (!cityData.getWorkingCity().equals("1")) {
+            Popup popup = new Popup();
+            popup.showCityNotAvailablePopupWindow(binding.tvCitySearch);
+        }
+        //city = String.valueOf(cityData.getId());
+        binding.tvCitySearch.setText(cityData.getCityName());
+        city = cityData.getId();
+    }
 }
