+++ app/src/main/java/com/dbot/client/main/profile/account/AccountInfoFragment.java
@@ -30,8 +30,8 @@ import com.dbot.client.common.SessionManager;
 import com.dbot.client.databinding.FragmentAccountInfoBinding;
 import com.dbot.client.login.LoginViewModel;
 import com.dbot.client.login.OtpEditText;
-import com.dbot.client.login.city.CityAdapter;
-import com.dbot.client.login.city.CityData;
+import com.dbot.client.common.city.CityAdapter;
+import com.dbot.client.common.city.CityData;
 import com.dbot.client.login.model.LoginResponse;
 import com.dbot.client.login.model.SignUpResponse;
 import com.dbot.client.login.model.User;
