+++ app/src/main/res/layout/fragment_home.xml
@@ -47,8 +47,21 @@
                         android:gravity="center"
                         android:text="Book your preferred slot:"
                         android:textSize="@dimen/_15ssp" />
-
-                    <Spinner
+                    <TextView
+                        android:id="@+id/tv_city_search"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_gravity="center"
+                        android:layout_marginStart="@dimen/_10sdp"
+                        android:layout_marginTop="@dimen/_10sdp"
+                        android:layout_marginEnd="@dimen/_10sdp"
+                        android:background="@drawable/bg_edit_text"
+                        android:drawableStart="@drawable/ic_location"
+                        android:drawableEnd="@drawable/ic_arrow_down"
+                        android:drawablePadding="10dp"
+                        android:gravity="start"
+                        android:textSize="@dimen/_12ssp" />
+                    <!--<Spinner
                         android:id="@+id/sp_city"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
@@ -56,7 +69,7 @@
                         android:layout_marginStart="@dimen/_10sdp"
                         android:layout_marginTop="@dimen/_10sdp"
                         android:layout_marginEnd="@dimen/_10sdp"
-                        android:gravity="center" />
+                        android:gravity="center" />-->
 
                     <TextView
                         android:id="@+id/tv_available_message"
