+++ app/src/main/res/layout/home2.xml
@@ -6,6 +6,11 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     tools:context=".main.home.HomeFragment">
+    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:id="@+id/swipeRefreshLayout2"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/ns_home2"
@@ -552,6 +557,6 @@
                 android:textSize="@dimen/_13ssp" />
         </LinearLayout>
     </androidx.core.widget.NestedScrollView>
-
+    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
